Renamed phase space module
